requirements.txt
dj-config-url==0.1.1
Django==1.11.4
olefile==0.44
Pillow==4.2.1
python-decouple==3.1
pytz==2017.2
whitenoise==3.3.0
